+++ curlio/quick.hpp
@@ -2,3 +2,4 @@
 
 #include "quick/form.hpp"
 #include "quick/json.hpp"
+#include "quick/reader.hpp"